requirements.txt
requests==2.31.0
feedparser>=6.0.11
openai>=2.7.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml>=5.0.0 